refactor: rename options page tab components
