+++ src/entrypoints/options/RuleTab/rule.tab.module.css
@@ -1,4 +1,4 @@
-main.rule-list-tab {
+main.rule-tab {
 	display: flex;
 	flex-direction: column;
